clean up auto-buy. Notable issues:
goodtrait items, if not selected, weren't filtered as the comments stated
in `30% chance for her to buy any good restore item.' a choice on an
empty newpool could ctd.
if pool is empty (e.g. if a char has no money) the shuffle(pool) can ctd.
(ctd notions based on code, not observation)

Signed-off-by: picobyte <pico0obyte@gmail.com>